requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
python-dotenv
pydantic
sentence-transformers
groq
